AI-Server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pymongo==4.6.1
langchain==0.0.335
langchain-openai==0.0.8
pydantic==2.4.2
openai==1.20.0